refactor!(M2006SimpleFeederSystem): Align to new naming convention
